+++ src/main/java/es/ucm/fdi/avisos/business/entity/PrioridadesAvisoEnum.java
@@ -1,11 +1,11 @@
-package es.ucm.fdi.anuncios.business.entity;
+package es.ucm.fdi.avisos.business.entity;
 
-public enum TipoAvisoEnum {
-	URL("Enlace"), ADJUNTO("Adjunto"), HTML("Contenido HTML");
+public enum PrioridadesAvisoEnum {
+	NORMAL("Normal"), IMPORTANTE("Importante");
 	
 	private String description;
 	
-	private TipoAvisoEnum(String description) {
+	private PrioridadesAvisoEnum(String description) {
 		this.description = description;
 	}
