+++ src/main/webapp/WEB-INF/spring/root-context.xml
@@ -41,7 +41,7 @@
 		<property name="packagesToScan">
 			<array>
 				<value>es.ucm.fdi.users.business.entity</value>
-				<value>es.ucm.fdi.anuncios.business.entity</value>
+				<value>es.ucm.fdi.avisos.business.entity</value>
 				<value>es.ucm.fdi.storage.business.entity</value>
 				<value>es.ucm.fdi.espacios.business.entity</value>
 			</array>
