+++ src/main/webapp/WEB-INF/spring/webcontext/portal-context.xml
@@ -24,18 +24,18 @@
 
 
 	<!-- Packages to search for @Service beans definitions -->
-	<context:component-scan base-package="es.ucm.fdi.anuncios.business.boundary" />
+	<context:component-scan base-package="es.ucm.fdi.avisos.boundary" />
 	<context:component-scan base-package="es.ucm.fdi.storage.business.boundary" />
 	<context:component-scan base-package="es.ucm.fdi.espacios.business.boundary" />
 
 
 	<!-- Packages to search for @Controller beans definitions -->
 	<context:component-scan base-package="es.ucm.fdi.portal.web" />
-	<context:component-scan base-package="es.ucm.fdi.portal.anuncios.web" />
+	<context:component-scan base-package="es.ucm.fdi.avisos.web" />
 	<context:component-scan base-package="es.ucm.fdi.espacios.web" />
 	<context:component-scan base-package="es.ucm.fdi.storage.web" />
 	
-	<context:component-scan base-package="es.ucm.fdi.anuncios.validation" />
+	<context:component-scan base-package="es.ucm.fdi.avisos.validation" />
 	<context:component-scan base-package="es.ucm.fdi.espacios.validation" />
 	
 	<!-- Packages to search for @Repository beans definitions -->
@@ -43,7 +43,7 @@
 		entity-manager-factory-ref="rootEntityManagerFactory"
 		transaction-manager-ref="rootTransactionManager" />
 		
-	<data:repositories base-package="es.ucm.fdi.anuncios.business.control" 
+	<data:repositories base-package="es.ucm.fdi.avisos.business.control" 
 		entity-manager-factory-ref="rootEntityManagerFactory"
 		transaction-manager-ref="rootTransactionManager" />
 		
@@ -77,32 +77,6 @@
 		class="org.springframework.context.support.ReloadableResourceBundleMessageSource"
 		p:basename="classpath:messages" />
 		
-<!-- 	<bean id="portalSessionFactory" -->
-<!-- 		class="org.springframework.orm.hibernate4.LocalSessionFactoryBean" -->
-<!-- 		p:dataSource-ref="portalDataSource" p:packagesToScan="es.ucm.fdi.anuncios, es.ucm.fdi.espacios" -->
-<!-- 		p:hibernateProperties-ref="hibernateProperties" /> -->
-
-	<!-- HSQL -->
-<!-- 	<bean id="portalDataSource" class="org.apache.commons.dbcp.BasicDataSource" -->
-<!-- 		destroy-method="close"> -->
-<!-- 		<property name="driverClassName" value="org.hsqldb.jdbcDriver" /> -->
-<!-- 		 <property name="url" value="jdbc:hsqldb:./db2;shutdown=true" />  -->
-<!-- 		<property name="username" value="sa" /> -->
-<!-- 		<property name="password" value="" /> -->
-<!-- 	</bean> -->
-	
-<!-- 	<util:properties id="hibernateProperties"> -->
-<!-- 		<prop key="hibernate.dialect">org.hibernate.dialect.HSQLDialect</prop> -->
-<!-- 		<prop key="hibernate.max_fetch_depth">3</prop> -->
-<!-- 		<prop key="hibernate.jdbc.fetch_size">50</prop> -->
-<!-- 		<prop key="hibernate.jdbc.batch_size">10</prop> -->
-<!-- 		<prop key="hibernate.show_sql">false</prop> -->
-<!-- 		<prop key="hibernate.hbm2ddl.auto">create-drop</prop> -->
-<!-- 		<prop key="hibernate.hbm2ddl.auto">update</prop> -->
-<!-- 	</util:properties> -->
-	<!-- FIN HSQL -->
-	
-	
 	<!-- PostgreSQL -->
 <!-- 	<bean class="java.net.URI" id="dbUrl"> -->
 <!-- 		<constructor-arg value="#{systemEnvironment['DATABASE_URL']}" /> -->
