+++ src/test/java/es/ucm/fdi/anuncios/controllerTest/test-DispatcherServlet-context.xml
@@ -15,7 +15,7 @@
     ">
 
 	<!-- <mvc:annotation-driven />
-	<context:component-scan base-package="es.ucm.fdi.anuncios" /> -->
+	<context:component-scan base-package="es.ucm.fdi.avisos" /> -->
  
 	<bean
 		class="org.springframework.web.servlet.view.InternalResourceViewResolver">
